[NFC] Ensure TargetOpcode::* instructions are compatible with guessInstructionProperties=0
rL162640 introduced CodeGenTarget::guessInstructionProperties. If a target
sets guessInstructionProperties=0 in its FootgtInstrInfo, tablegen will error
if it has to guess properties from patterns. Unfortunately, this can't be used
with current upstream LLVM as instructions in the TargetOpcode namespace are
always included and have inferred properties for mayLoad and mayStore. This
patch provides the simplest possible fix to this problem, defaulting mayLoad
and mayStore in the TargetOpcode scope. The local definitions take precedence,
so definitions like G_LOAD or PatchPoint have their expected properties and
there is no functional change. This patch also adds hasSideEffects definitions
to those instructions where it is currently inferred.

Differential Revision: https://reviews.llvm.org/D37065